% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckner.R
\name{buckner17}
\alias{buckner17}
\title{Buckner 17-Network Cerebellar Atlas}
\usage{
buckner17()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cerebellar).
}
\description{
Cerebellar parcellation into Yeo 17 resting-state functional
networks from Buckner et al. (2011).
}
\examples{
buckner17()
plot(buckner17())
}
\references{
Buckner RL, et al. (2011). The organization of the
human cerebellum estimated by intrinsic functional
connectivity. \emph{Journal of Neurophysiology}, 106(5):2322-2345.
\doi{10.1152/jn.00339.2011}
}
\seealso{
Other ggseg_atlases: 
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{mdtb10}()},
\code{\link{nettekoven32}()},
\code{\link{xue10}()}

Other cerebellar_atlases: 
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{mdtb10}()},
\code{\link{nettekoven32}()},
\code{\link{xue10}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
