% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegCerebellum-package.R
\docType{package}
\name{ggsegCerebellum-package}
\alias{ggsegCerebellum}
\alias{ggsegCerebellum-package}
\title{ggsegCerebellum: Cerebellar Atlases for the 'ggsegverse' Ecosystem}
\description{
Collection of cerebellar parcellation atlases for the 'ggsegverse' ecosystem. Includes anatomical (Diedrichsen 2009), resting-state network (Buckner 2011, Ji 2019, Xue 2021), task-based (King/MDTB 2019), and hierarchical functional (Nettekoven 2024) parcellations. All atlases derived from the DiedrichsenLab cerebellar atlas collection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegCerebellum}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegCerebellum/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
