% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji.R
\name{ji10}
\alias{ji10}
\title{Ji 10-Network Cerebellar Atlas}
\usage{
ji10()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cerebellar).
}
\description{
Cerebellar parcellation into 10 subcortical functional
networks from Ji et al. (2019).
}
\examples{
ji10()
plot(ji10())
}
\references{
Ji JL, et al. (2019). Mapping the human brain's
cortical-subcortical functional network organization.
\emph{NeuroImage}, 185:35-57.
\doi{10.1016/j.neuroimage.2018.10.006}
}
\seealso{
Other ggseg_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{mdtb10}()},
\code{\link{nettekoven32}()},
\code{\link{xue10}()}

Other cerebellar_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{mdtb10}()},
\code{\link{nettekoven32}()},
\code{\link{xue10}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
