% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtb.R
\name{mdtb10}
\alias{mdtb10}
\title{MDTB 10-Region Cerebellar Atlas}
\usage{
mdtb10()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cerebellar).
}
\description{
Multi-Domain Task Battery cerebellar parcellation with 10
task-based functional regions from King et al. (2019).
}
\examples{
mdtb10()
plot(mdtb10())
}
\references{
King M, et al. (2019). Functional boundaries in
the human cerebellum revealed by a multi-domain task
battery. \emph{Nature Neuroscience}, 22(8):1371-1378.
\doi{10.1038/s41593-019-0436-x}
}
\seealso{
Other ggseg_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{nettekoven32}()},
\code{\link{xue10}()}

Other cerebellar_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{nettekoven32}()},
\code{\link{xue10}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
