% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettekoven.R
\name{nettekoven32}
\alias{nettekoven32}
\alias{nettekoven68}
\title{Nettekoven Cerebellar Functional Atlases}
\usage{
nettekoven32()

nettekoven68()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cerebellar).
}
\description{
Hierarchical symmetric cerebellar parcellations from
Nettekoven et al. (2024). Available as 32-region and
68-region variants. Regions are coded by functional domain:
Motor (M), Auditory (A), Default mode (D), Somatosensory (S),
with left (L) and right (R) hemispheric labels.
}
\examples{
nettekoven32()
plot(nettekoven32())
nettekoven68()
plot(nettekoven68())
}
\references{
Nettekoven C, et al. (2024). A hierarchical atlas
of the human cerebellum for functional precision mapping.
\emph{Nature Communications}, 15:8376.
\doi{10.1038/s41467-024-52371-w}
}
\seealso{
Other ggseg_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{mdtb10}()},
\code{\link{xue10}()}

Other cerebellar_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{mdtb10}()},
\code{\link{xue10}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
