% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xue.R
\name{xue10}
\alias{xue10}
\title{Xue 10-Network Cerebellar Atlas}
\usage{
xue10()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cerebellar).
}
\description{
Individual resting-state cerebellar parcellation with 10
networks from Xue et al. (2021).
}
\examples{
xue10()
plot(xue10())
}
\references{
Xue A, et al. (2021). Local-global parcellation
of the human cerebellum from intrinsic functional
connectivity. \emph{Cerebral Cortex}, 31(2):764-778.
\doi{10.1093/cercor/bhaa253}
}
\seealso{
Other ggseg_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{mdtb10}()},
\code{\link{nettekoven32}()}

Other cerebellar_atlases: 
\code{\link{buckner17}()},
\code{\link{buckner7}()},
\code{\link{ji10}()},
\code{\link{mdtb10}()},
\code{\link{nettekoven32}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
